+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fix alias list overwrite during profile update.
+- Require alias to be different account.
 
 ## [2.0.0] - 2023-10-02